Remove uneeded vehicle/const.py
